diskdump: fix handling of max_mapnr
The use of max_mapnr is broken. It is first used to determine whether
two bitmaps are present (so it must be valid), and then it is reset
to the total number of bits in the bitmap.

If this was used to deal with overflow of the 32-bit counter in the
header, then it is the wrong approach. Instead, such big machines
should always use a version of makedumpfile which stores the full
64-bit value in the subheader.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>