Add process_load to arch_ops
This can be used to add an architecture-specific hook to processing
of ELF program headers. The x86_64 architecture can use this hook to
determine phys_base.

Note that only a small subset of the program header fields are passed
to this hook. Only as much as is needed to determine phys_base on
x86_64. Since this API is internal to the library, it can be extended
if more values are needed for something, but passing down tons of
arguments just for the sake of clean APIs is not worth the effort.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>