x86_64: determine phys_base from a kernel text ELF LOAD segment
This logic is slightly different from the code in crash or makedumpfile.
Basically, the conditional in x86_64_process_load() filters out only
segments that contain some of the kernel text+static data.

The calculation takes the difference between the actual physical
address where the kernel is loaded and the "preferred" load address
(i.e. without relocation).

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>