Set DIF_UTSNAME when utsname is complete
Set a flag when utsname is completely initialised. Note that it is
possible that only some fields in utsname are initialised (because
they are determined indirectly, e.g. from VMCOREINFO). In that case,
the flag should not be set, so the generic code can attempt an
alternative approach to fill in _all_ fields.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>