Provide a generic interface for using virtual addresses
Virtual address translation is needed to read values from the crashed
kernel, which is in turn sometimes needed to initialise fields that are
otherwise stored in the file header. This could be done by users of
the library, but that would be code duplication. Plus, it is convenient
if the library does it automatically.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>