Remove DIF_PHYS_BASE check from x86_64_vtop()
Many dump files do not provide any way to determine the correct phys_base
offset, but the vast majority of them in fact uses an offset of zero.
It's probably better to return potentially incorrect data (and validate
it afterwards) than returning nothing and giving the caller no way to
get the data.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>