x86_64: Set up kernel virtual memory layout
The virtual memory layout for x86_64 has changed several times. However,
during late_init, the kernel version is already known (if it can be
determined), so the appropriate map can be chosen.

Note that if the kernel version cannot be determined, version_code
is guaranteed to be zero, and the "stub" layout is kept.

The layouts are based on Documentation/x86/x86_64/mm.txt in the Linux
kernel sources.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>