Remove the provisionary virtual memory mappings in kdump_open()
In most cases, these mappings are bogus and may only confuse the
library user, because kdump_vtop() does not fail where it should.
The provisionary mapping is only good enough to read kernel text/data
as part of the initialisation, and it should not be used afterwards.

This commit makes sure that it won't be used by mistake.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>