x86_64: Use fixed page size for internal processing
Basic page size is constant on x86_64, so let's use that constant
instead of any value obtained from the dump file.

I would add a sanity check, but that would make libkdumpfile unusable
on dump files where the stored page size is incorrect. While such
dump files are probably corrupted, the stored page size is not needed
on x86_64, and some useful information may still be obtained from such
files.

I did not go as far as changing ctx->page_size, as it may be displayed
to the end users, making it easier to spot that there is something wrong
with the dump file.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>