Set default page shift in kdump_set_arch()
If page size is still unknown when architecture is being set, the default
for that architecture should be used.

Note that this currently works only for architectures with a fixed page
size. Where multiple page sizes are possible, setting the architecture
fails. I'm not sure if silently using a default value is better...

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>