Separate processing of noarch and arch notes
With ELF notes, there is a chicken-and-egg problem: some notes can be
processed only after arch-dependent fields are initialized (e.g. page
size), but those fields are initialised by processing ELF notes.

To solve this, the notes must be processed twice. First, the
arch-independent notes are processed, and then the arch-dependent ones.

This commit adds the function calls to process arch-dependent and
arch-independent notes separately.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>