Implement vtop for ia32 (non-PAE)
Add code to walk ia32 non-PAE pagetables. Same restrictions as with PAE.

Note the detection of PAE. Top of virtual memory is always mapped
(it contains fixmaps), so entry 3 in PGD must always be initialised
in PAE mode. But if the top-level table is in fact non-PAE, then the
same area corresponds to slots 6 and 7 of the PGD, which maps virtual
address range 0x1800000-0x1ffffff. These addresses are never mapped
in swapper_pg_dir, so they are left uninitialized (zero).

In other words, if bytes at offset 24 to 31 from swapper_pg_dir are
all zero, it must be a non-PAE page directory, and if they are non-zero,
it must be a PAE page directory.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>