Use VMCOREINFO to detect ia32 PAE
This method should be faster, but it is insufficient: If the CONFIG
row is found, then the kernel must have used PAE, but if it is not
found, it could either be a non-PAE kernel, or VMCOREINFO is missing.

Since most 32-bit kernel dumps are PAE, let's do this optimisation
anyway.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>