Cleanup: rename all internal symbols
Global symbols that are only used internally by the library should
have a name that minimizes the risk of clashes with the application.
This is not an issue for dynamic link (where the linker map file takes
care of hiding these symbols), but it may happen in a static link.

Use a preprocessor macro to do the rename, so that the (ugly) real
names are not seen in the source files.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>