kdump_set_fd: stop on genuine errors
If a probe fails with an error code other than kdump_unsupported, it
means the file format is recognized, but something failed during
initialization. Probing another format is unlikely to succeed, but
it would clear the actual error message, making it harder to understand
what failed. Let's stop the search in that case and keep the error
message from the failed probe.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>