Improve error reporting of paged reads
The paged_cpin interface is suboptimal, as it only returns the number
of remaining bytes. While this makes it very easy to check if the
complete buffer was read (zero return value), it's impossible to
distinguish between a system error and EOF if the return value is
non-zero. As a consequence, EOF may be reported as "Success", which
is wrong and confusing.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>