Fix MFN translation for DomU dumps
Pages in a DomU dump are saved in an arbitrary order. To find a given
MFN, it must be found in the .xen_p2m array.

Note that this array does not exist if physmap is auto-translated (e.g.
hvm domains on x86_64). Store this information in kdump_ctx (as
xen_pte_is_mach) to make translation of virtual addresses possible.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>