Look up global templates by their GKI_ index
Speed up global template look-up by using an index. This is implemented
as special pointer values near the end of addressable range. These values
are not usable by Linux applications anyway, because they are also used
as error values by the kernel.

With this commit, it is possible to use GATTR(GKI_linux_uts_sysname)
instead of "linux.uts.sysname". The latter is still possible, but the
key is searched in a loop, whereas the former returns the key directly.
By using a pre-defined constant, the GATTR() is also less prone to typos.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>