Export kdump_get_string_attr
This is a handy shortcut. However, keep in mind that it does not provide
full diagnostics on failure:

 - The exact return value of kdump_get_attr() is lost.
 - A static string may contain the value NULL (e.g. a VMCOREINFO line
   without an equal sign).

You can still obtain the error string with kdump_err_str() and check
if it is NULL, but then it's probably easier to use kdump_get_attr()
directly.

In short, this function is good if the return value can't be sensibly
NULL, and you don't need exact diagnostics.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>