Remove prev pointer from struct attr_data
This field is used only when deleting attributes. If it is missing, all
children of the parent node must be enumerated to find the back link.
Since the delete operation is relatively rare, the savings are not worth
an extra field.

For an x86_64 system with 1024 CPUs, this change currently saves 28 KiB
at run time. It will be more if more registers can be parsed (e.g. QEMU
dumps).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>