Fix make_attr_path for the root node
Root node should be denoted by an empty string, so make_attr_path() must
put the NUL character into the buffer.

While fixing this bug, let's also adjust the end pointer to match the
description (let it point to the end of the buffer, not one byte beyond
the end).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>