Create an unset attribute for each directory template
This means that all potential attributes are pre-allocated when adding
the attribute template. Global attributes are allocated at initialization,
dynamic attributes are allocated with the template.

This approach greatly simplifies instantiate_path(). Since all attributes
in the hierarchy must be already allocated (just not set), instantiating
a path becomes simply setting isset to 1 for all directories up the
hierarchy.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>