Set xen.phys_start attribute if available
The format of the Xen crash info note has changed several times, adding
fields to the end and (sic!) to the middle. The physical base address
is available on x86_64 and ia32 iff the core file was generated with
a Xen version that includes changeset 439a3e9459f2.

It is important to leave the attribute uninitialized, so library users
have a chance to provide their own value.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>