Provide an arch op for mfn-to-pfn translation
Why an arch op if there is a format op already? Well, the format op is
good if the information is stored in the dump file itself (i.e. xc_core
style dumps). The arch op is necessary if the information must be
retrieved from the raw data itself using an arch-specific method (i.e.
system dumps).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>