Provide virtual-to-machine translation
For a non-auto-translated Xen domain, the machine address can be obtained
directly from the page tables (unlike the physical address, which must be
mapped back from the physical address first).

So, if the caller wants the machine address, libkdumpfile will first
translate the virtual address to a machine address, then look up the
physical address for that machine address, and then the caller would
translate this physical address back to a machine address.

Obviously, it's much faster to avoid the double-translation. ;-)

Signed-off-by: Petr Tesarik <ptesarik@suse.com>