Set phys_base from vtop_init() if necessary and possible
If the physical base was not stored in the dump file, try to determine its
value from the page tables.

However, if the physical base cannot be determined this way, do not fail the
initialization; just convert all XLAT_KTEXT entries into XLAT_VTOP, which is
slower, but can always be tried.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>