Replace read flags parameter with address space
These are no "flags", e.g. only one of them can be specified. In fact,
it simply says how the address argument should be interpreted.

With this change, keeping the parameter as the last argument to
kdump_read() is illogical; it determines the type of the addr
parameter, so they should be grouped together.

Note that this change breaks source and binary compatibility with
older library versions, but hopefully there aren't many users yet,
so let's clean up the API now (before it's too late).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>