Clean up the semantics of read_page format op
There is some confusion about the meaning of the parameter to read_page().
For most files, it is a machine physical address (requiring additional
translation for Xen Dom0), but for xc-generated DomU dumps, it is the
kernel physical frame number, and MFNs must be read with read_xenmach_page().

Why?

The read_page() op should always read the machine physical address. This
is sufficient for any file format, but may be suboptimal for Xen PV dumps,
because the PFN or MFN must be translated to an index inside .xen_pages.
The p2m translation uses the same array, so it would be scanned twice for
no good reason.

For such files, let the format handler provide a shortcut with read_kpage().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>