Store vtop_pgt function in struct vtop_map and stop using indices
With this change in place, struct vtop_map contains all the information
needed to translate a virtual address to a physical address. As a result,
a pointer to this struct can be passed to map_vtop instead of a map
index and the array can be replaced with two struct instances.

Less typing in the usual case.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>