LKCD: Do not rely on max_pfn
It appears that max_pfn does not specify the maximum PFN in the file, so
it cannot be trusted. It is not checked in get_page_desc() anyway, so if
a file contains a page beyond max_pfn, get_pfn_slot() may try to use an
index beyond the size of pfn_level1[].

Solve this by resizing the level1 table dynamically as needed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>