Test suite: Fix a few Automake issues
First, automake warns that "testutil.a is not a standard library name".
While I couldn't care less as long as it compiles, it might actually
do some harm on other architectures, so let's add the "lib" prefix.

Second, using LIBS does not really work very well, as the library is
not added to dependencies.

Third, the global LDADD must be added to checkattr_LDADD explicitly.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>