LKCD: Allow gaps in PFN blocks
It is common that only a few pages are missing from an otherwise
continuous PFN block. The current code will allocate a separate block
in that case, but that is suboptimal:

  1. The PFN block chain grows, slowing down lookup.
  2. Each PFN block header occupies some space, so it is even less
     efficient for memory requirements.

Instead, let's use a special value to mark that the PFN is missing
(is a "gap"). I can use zero, because it corresponds to the first
PFN in the block, which is not stored (thanks to commit 0b4f233f1ceb).

Before this commit, the most common length of a PFN block chain is
between 200 and 250. After this commit, almost 70% chains are shorter
than 10 links.

Note that gaps may be filled by pages found later in the dump file.
However, since only 32 bits of the offset are stored in the offs
array (to conserve memory), it may happen that the offset cannot be
represented. In that case, the existing PFN block must be split into
two and a new PFN block must be inserted in the middle. The
lkcd-unordered-faroff test has been extended to check that the
split is done correctly.

Also, the lookup function had to be adjusted, so that it can find
blocks that can be extended up to the current PFN while searching
the dump file for pages.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>