LKCD: Return an error if offs cannot be allocated
The idea is that a new block can be tried instead of failing.

But it does not work properly:

1. block->n had already been set, and
2. when adding in the middle of the block, the block might need to
   be split.

Since it is quite unlikely that a subsequent PFN block allocation would
be successful, let's just fail.

The original intention was to cope with virtuall address space
fragmentation (e.g. on 32-bit systems) when allocating huge arrays.
This is less of an issue when PFN blocks cannot cross a level-3
boundary, because the array size is capped at PFN_IDX3_SIZE.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>