Store max_pfn as an attribute
Make the max_pfn value accessible by library users. This commit also
clarifies the meaning of max_pfn. It is in fact one higher than the
maximum valid PFN for the given dump file. Confusing, as it may sound,
this definition is consistent with Linux kernel's max_pfn variable.

Note that this value is not available for LKCD dumps. Previously,
LKCD dumps used dh_memory_size for this, but that's wrong, because
dh_memory_size corresponds to totalram_pages (i.e. excluding any
memory holes).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>