cleanup: Fix confusing attribute hash table implementation
The first hash table is embedded inside kdump_ctx. This is not a big
problem, but it means that lookups cannot cleanly return a writeable
object from a const object. Usually, this is not a problem, but it
requires occasional ugly typecasts and does not bring much in return
(saving one pointer dereference).

Simplify the logic and get rid of the embedded hash table. Note that
this change may improve cacheline alignment of the first hash table,
possibly compensating for the extra dereference (not measured!).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>