Fix a warning about redefining _POSIX_C_SOURCE
This feature-test macro must be defined before any standard header
files are included, otherwise it has already been set to a default
value.

This is in fact an issue with the Python header file, which shouldn't
try to define this macro, but including it as the first file can
work around the issue.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>