ELF: Deal with fractional pages
As found in the makedumpfile project the other day, ELF dump may
contain incomplete pages. It seems that such a fractional page can
always be found as a complete page elsewhere in the ELF dump, so it
doesn't have to be combined from two sources. However, a check for
the page end must still be added to avoid incorrect read of the
fractional page.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>