cache: Store in-flight cache entries in a list
This is the first step towards parallel processing. If an entry
is already in-flight, return this entry to prevent duplicates.

There is no way to distinguish an in-flight entry from a cache miss,
but such distinction would be of little use to the caller, because
it (presumably) cannot continue without the data content, anyway.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>