cache: Slightly change the replacement policy on cache miss
The purpose of nprobetotal is to avoid wiping all cache history during
a sequential scan. However, the total size of the probe list may be
bigger than half the cache, e.g. during an initial fill, there's no
better use for the second half. In other words, the condition should
be checked only if the cache is already full (most of the time).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>