ELF: recreate partial pages
The logic follows that of makedumpfile:

  - fill holes with zeroes (see Note below),
  - fill p_filesz portions with file data and
  - fill remaining p_memsz portions again with zeroes.

Note: A "hole" is data at a physical address that is not covered by any
ELF LOAD program header. In other words, the ELF file does not specify
any data for such a hole (not even zeroes). So, why does libkdumpfile
fill them with zeroes? It's because the library works with page
granularity (the compressed format does not even have a way to store
a partial page), so if only part of a page is stored, a complete page
must be provided to make this partial data accessible.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>