cache_get_entry: Return NULL if all entries are used up
Since the cache can be used recursively, the cache can run out of data
entries if the recursion level is too high.

This condition may have to be propagated up to the library user, but
note that it can never happen right now, because the kdump object
itself is not re-entrant. However, the code is now prepared for future
enhancements.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>