Make cache thread-safe
Add a mutex to guard critical sections. Note that since decrementing the
reference counter is not atomic, the mutex must also be used by
cache_put_entry(), requiring the use of the "cache" parameter.

This is sufficient to make multi-threaded ELF reads pass, but note that
most operations are still not thread-safe. Even reading non-ELF dumps
is not thread-safe...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>