python: Define a custom type for attribute directory
Copying the complete attribute tree to Python dicts is suboptimal. Not only
does it waste some memory, but the copy does not change if some of the
kdumpfile attributes are changed. The only way would be to make a new
complete copy of the hierarchy.

Let's do this differently and define our own type for this.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>