python: Allow lookup as Python attributes
Allow syntax such as "attr.arch.name", which is equivallent to
attr['arch']['name'] or attr['arch.name'].

This is a bit tricky, as we don't want to override the default
mechanism completely. So, first try the generic method. If that
fails, try to lookup the libkdumpfile attribute.

Last but not least, if the libkdumpfile attribute is not found,
then raise an AttributeError exception, because this one is often
handled by callers (unlike the kdumpfile-specific types).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>