Fix lookup_attr_part()
While writing keycmp_dir(), I realized that the regular keycmp() does not
check whether the last found parent is the root node. So, in fact, this
has always been buggy, but the bug has never been found, because a hash
collision with a matching subpath is quite unlikely.

This wipes off any potential savings of lookup_attr_part(), compared to
lookup_attr_dir(), so let's merge the two.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>