python: Raise KeyError when accessing an unknown key
When trying to access a non-existent attribute using attr['key'], raise
KeyError instead of the kdumpfile-specific exceptions. This makes the
attribute-directory object behave more like a dictionary. This is generally
a good thing, because standard Python exceptions are already known to
newcomers, so there are less new things to learn.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>