Do not fail if if architecture is unknown
KDUMP files produced by very old versions of makedumpfile did not fill
utsname fields in the header, so architecture cannot be determined.
It is still possible to read the raw data and guess; this is done e.g.
by kdumpid.

This solution changes the implementation so that architecture gets
set automatically when uts.machine is set. If architecture is not set,
then architecture-specific initialization is skipped.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>