optimize arch_name()
Convert an array of pointers to an array of strings. The resulting
table does not need any dynamic relocations, so it can be put into
the read-only data section (and shared between library users).

Note: Specifying the terminating NUL character explicitly results
in a compiler warning if any of the arch names are longer than
MAX_ARCH_NAME_LEN.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>