Rename x86/i386 to ia32
This is consistent with the source file naming, and it is in fact the
official Intel name for the architecture.

Note that since "ia32" is not a valid machine name, but arch.name
post-set hook uses that function, arch-specific initialization is
broken for ia32, and most i386 tests will fail until this is fixed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>