Centralize attribute validation
This function is currently just a wrapper around attr_isset(), but the
semantics is different:

  - attr_isset() simply checks whether the attribute value is set
  - validate_attr() may take an action to update the value

There is currently no method to set the value, so there is no
practical difference. Note, however, that validate_attr() already
takes a non-const pointer to the attribute, which needs some
further adjustments in the callers.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>