Implement attribute overrides
A format handler or arch handler may need to override the operations for
certain attributes. This can be done by modifying the template pointer,
but it can become quite messy if every handler open-codes the algorithm.

Instead, provide a universal method for making and undoing such custom
modifications. It is called "overrides", because the user can use struct
attr_override to override some aspects of the default template.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>