lkcd: Get max_pfn attribute on demand
For LKCD dump files, the maximum pfn is not stored in the header, and
because of the file format, it cannot be determined without reading the
complete file. This is slow if max_pfn is not needed by the caller.

Here, the issue is solved by installing a validate op for the max_pfn
attribute and reading the file only if the value of max_pfn is actually
requested by the library user.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>