Allocate separate compressed buffers
The current buffer handling is hackish. During file open, a buffer is
allocated for the file headers (so that each probe does not have to
read it on its own). This buffer may be later reused for compressed
data (LKCD and diskdump formats only).

This breaks s390x LKCD dumps if VMCOREINFO is read from the dump itself
as part of arch initialization, because uts fields are set from the
header (stored in the scratch buffer), which may get overwritten by
lkcd_read_page().

Signed-off-by: Petr Tesarik <ptesarik@suse.com>