Prefer non-empty uts fields
Some file formats store uts fields in the header, but producers fail to
fill them in. This is not a problem in itself (just that the data is
unavailable), but it may be an issue if there is an alternative way to
find the actual uts fields, but the value is later overwritten with the
(uninitialized) value from the header.

To solve it, do not change the value to an empty string if there is
already a valid non-empty value.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>