Add cache reference counting
This change is necessary to avoid locking the whole kdump_shared
structure while reading a page. The code in def_read_cache() simply
takes a reference to the current cache and stores it in struct page_io.

If another thread reallocates shared->cache after that, the original
thread can still complete the operation with the current cache object,
because it holds an extra reference to that object.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>