Add proper locking to kdump_get_attr special shorthands
All the shorthand calls must also acquire the shared lock for reading
to ensure consistency.

This is the safe and portable but heavy-weight method. Many computer
architectures do not need any locking here, but some do. Most notably:

  a. If kdump_value_t cannot be read with a single instruction.
  b. If special care is needed to ensure cache coherency (e.g.
     Intel Itanium Platform).

The locking overhead is rather big and not needed on the most common
architectures. OTOH these calls are presumable not too frequent.
If the locking proves to be a bottleneck for any workload, there is
some room for improvement here.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>