Make lkcd_max_pfn_validate thread-safe
It is important to understand why it is safe to modify the target
attribute in validate(), although the shared lock is only held for
reading. If the attribute is not yet valid, it does not have any
previous value, so another thread can't be possibly reading the
value while it is being changed. Multiple readers may still enter
the validate() function at the same time; but this is guarded by
pfn_block_mutex.

Note that it is not safe to remove the override, because this
touches other overrides in the chain, which is racy. However, it
is safe to modify lkcd own override while it is protected by
pfn_block_mutex.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>