Distinguish between persistent and volatile attributes
Persistent attributes will survive a clear_attr(). This is important
when probing a file format. A format handler may start seting additional
attributes, but these attributes should be removed again if the probe
fails in the end.

All attributes that are set explicitly by the user are persistent.
Most other attributes are volatile, but for example cache.size is not,
because the default cache size is still valid after a probe failes.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>