Pass attribute flags to set_attr()
Replace the persist parameter with the complete flags, which should
allow to specify more than one flag in a single parameter.

Note the definition of ATTR_DEFAULT and ATTR_PERSIST. The use of
inline functions allows gcc to optimize the macro to a single register
load. I haven't found a simpler way to achieve it.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>