addrxlat: store raw PTE value in addrxlat_vtop_state
The value may be needed for verbose VTOP reporting. However, if the
target is live, the caller may have no way of getting the value that
was used by libaddrxlat, because it may have changed meanwhile.

Storing this value in the state also allows to move the PTE read
outside arch-specific handlers, making the code a bit more DRY.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>