addrxlat: Use an addrxlat_pgt_t for ADDRXLAT_PGT method
The base address itself is not sufficient to walk the page table hierarchy.
However, with the previous commit in place, the addrxlat_pgt_t struct has
all the necessary information.

Note that ADDRXLAT_PGT_IND no longer makes sense. ADDRXLAT_PGT is already
indirect, because it does not contains a copy of the pagetable translation
object but only a reference to it.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>