addrxlat: Move pgt pointer away from addrxlat_ctx
The context is not necessarily bound to a single page table translation
object. The context should contain only data specific to a specific user,
such as callbacks or error message strings. Moreover, the context may be
useful even without page table translation.

This also means that kdumpfile's shared->pgtxlat is no longer used for
anything, and it can be removed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>