tests/xlat-os: Use pgtroot definitions
Set the root page table translation object before calling
addrxlat_osmap_init(). This way it can already be used by the
arch-specific initialization code if needed.

Note that the translation definition is incomplete when passed to
addrxlat_osmap_set_xlat(); it does not have any paging form. This
should be handled in a better way, but I have no better ideas,
and the current solution works...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>