addrxlat: Implement OS-specific init for Xen on x86_64
This is mostly new code, based on analysis of the xen git repository
and various changes in virtual memory layout.

The test suite covers all variants, but it should be improved by
including omitted mappings (which might be mistaken for a different
type of mapping). See xlat-xen-x86_64-4.6-bigmem.data for an example.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>