diskdump: Optimize page_desc mapping
The pfn_to_pdpos() method turns out to be a bottleneck on large
dumps. That's because the whole bitmap array must be scanned for
every page read. Instead, store the mapping from a PFN to a page
descriptor file position.

Now, this alone would increase memory requirements by a factor of
almost 200. However, real dumps tend to contain clusters of present
and excluded pages, so this information can be compressed quite
nicely. The only drawback is that the resulting array cannot be
indexed by PFN directly, but instead sorted and searched by interval
halving.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>