addrxlat: Rename pte_ppc64 to pte_ppc64_linux and remove checks
The page table translation implemented in libkdumpfile/libaddrxlat
is defined by software (here Linux), so it should not use a generic
identifier (such as addrxlat_pte_ppc64).

Next, the size check in walk_init_ppc64 is wrong. The Linux kernel
never performs the walk for addresses outside PGTABLE_RANGE, but it
is theoretically possible to set up a page table hierarchy for any
effective address, if needed. In other words, there is nothing to be
checked in a walk init function, because all effective addresses
are potentially valid. Instead, the caller should use the high bits
of the address to choose the right root page table base (possibly
none).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>