Rename addrxlat_def_* to addrxlat_meth_*
A "definition" is better suited for the parameters that define the
actual translation method, rather than the method itself, but I
did not have this idea when renaming the functions for the first
time.

Also, the functions are moved to meth.c, because the old name has
been misleading ever since generalizing the translation routines.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>