addrxlat: Unify address translation method definitions
Clean up the interface to addrxlat_meth_t. Previously, the kind of
translation was set implicitly with a call to one of:

- addrxlat_meth_set_none
- addrxlat_meth_set_offset
- addrxlat_meth_set_form

Funnily, addrxlat_meth_set_root() did not change the kind, although
it only made sense with ADDRXLAT_PGT.

Also, the corresponding addrxlat_meth_get_* functions did not check
the kind, and in fact didn't allow to return an error code if it did
not match the type of information requested.

Note that calling the wrong function actually violated the C language
specification, as it is undefined behaviour if you try to read a
different field of an enum than the last written one.

After the change there are only two functions:

- addrxlat_meth_set_def
- addrxlat_meth_get_def

And there is a new structure (addrxlat_def_t) which defines both
the kind and any additional parameters (such as paging form, root
page table address, or linear offset).

This a large commit, because it changes the API, so all users must
be updated, but it should have zero impact on the behaviour.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>