tests: Fix unescaping in the parameter parser
Last character of the escape sequence is repeated in the output
(e.g. "\n" becomes "\nn").

As a side note, the argument needs not be set to NULL if a NUL
character is found after the backslash, since there is no error
handling anyway...

Found while smoke-testing the addrxlat option parser.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>