addrxlat: Implement table lookup method
The ADDRXLAT_LOOKUP method uses a table to perform the translation.
It would be possible to achieve the same result with multiple
ADDRCLAT_LINEAR methods and map entries, but:

- Using a separate translation method allows to modify the
  translation without modifying the map itself.
- The data structures are more space efficient than allocating
  a separate translation method for each object.
- Adding more entries in the map does not scale, because map
  lookup is sequential.

Yes, the last argument is weak, because the map algorithm could be
changed (although not without adding more fields to the map data
structure).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>