addrxlat: x86_64: Fix Linux ktext automatic setup
Linux kernel text mapping on x86_64 is wrong in many ways, mainly
because setting it up is challenging - the kernel may be relocated,
and there are several ways to calculate the correct offset.

The new logic is (in descending order of priority):

1. If an ADDRXLAT_OSMAP_KTEXT method is already set before
   calling addrxlat_osmap_init(), then it is used.

2. If the physical base is specified with a physbase= option,
   it is used. This allows overriding whatever internal logic.

3. If a usable ADDRXLAT_OSMAP_PGT method is available, then
   the physical base address is calculated by walking the page
   tables.

Additionally, the physical offset no longer defaults to zero. If
it cannot be determined, page table translation is used for the
Linux kernel text region.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>