addrxlat: Specify target address space in translation method
Currently, the target address space is hard-coded in
addrxlat_walk_next() to ADDRXLAT_KPHYSADDR. This does not work
very well, because the same method is used for page table walks
(where the target is usually a machine physical address) and for
linear mappings (where the target is usually a kernel physical
address).

Instead of adding another translation kind, specify the target
address space in addrxlat_def_t.

This address space is now also used as the default address space
for PTE lookups, but the walk function can override it (see e.g.
the code for PPC64).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>