addrxlat/opt: Improve alignment of option table entries
The problem with a flexible array in an array is that the
element size is not aligned at all, so the second member of
an option array may end up on an odd address.

If you declare a real struct, this lets the compiler choose
a better alignment. On the other hand, this alignment is
implementation-dependent, so the element size must also be
stored in the main options array.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>