addrxlat: Generalize the LOOKUP method
This method can be useful for any translation based on fixed tables,
not just a virtual-to-physical translation.

The fields in addrxlat_lookup_elem_t have been renamed accordingly,
and the walk init function now uses the target address space from
the method definition.

Note that this has uncovered a bug in ppc64 translation system
initialization. The translation itself was OK, but only because
the walk init function ignored meth->def.target_as.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>