addrxlat_map_set: Allow clearing a map entry
This function could be used to unset a map range by setting the
method to NULL. Unfortunately, this segfaults, because
addrxlat_map_set() tries to increase the reference counter,
dereferencing a NULL pointer.

An interesting corner case is clearing an entry in an
uninitialized map. The code has assumed that a new entry cannot
merge with anything in an empty map, but a NULL entry actually
merges with the whole range, so this must be checked explicitly.
Note that addrxlat_map_set() cannot simply return NULL in this
case, because a NULL return value is reserved for allocation error.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>