addrxlat: Pass full address to addrxlat_by_map()
The current API has a serious flaw: if a map contains multiple
methods with different target address spaces, there is no way to
let the caller know to which address space the result was translated.

This can be fixed by using a full address, but such change ripples
through the code. Let's chew the bullet now.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>